Add files via upload
The dataset can be used as a basis to select players for the upcoming FIFA across the globe by focusing on the below objectives:
1. Is the potential of a player dependent on his age?
2. To create the best model for predicting the potential of a player based on several attributes.
We used Linear Regression to find whether the potential of a player is dependent on his age and other skill sets or not. For predicting the best model among we have created various models varying from Regression to Auto neural and then fed the outputs to Model Comparison. This model comparison will fulfil our second objective.